improve code, cleanup
