+++ cli/src/main/java/com/devonfw/tools/ide/commandlet/CreateCommandlet.java
@@ -5,7 +5,6 @@ import java.nio.file.Path;
 
 import com.devonfw.tools.ide.context.IdeContext;
 import com.devonfw.tools.ide.git.GitUrl;
-import com.devonfw.tools.ide.git.repository.RepositoryCommandlet;
 import com.devonfw.tools.ide.io.FileAccess;
 import com.devonfw.tools.ide.property.FlagProperty;
 import com.devonfw.tools.ide.property.StringProperty;
@@ -17,9 +16,6 @@ public class CreateCommandlet extends AbstractUpdateCommandlet {
 
   /** {@link StringProperty} for the name of the new project */
   public final StringProperty newProject;
-
-  /** {@link FlagProperty} for skipping the setup of git repositories */
-  public final FlagProperty skipRepositories;
 
   /** {@link FlagProperty} for creating a project with settings inside a code repository */
   public final FlagProperty codeRepositoryFlag;
@@ -33,7 +29,6 @@ public class CreateCommandlet extends AbstractUpdateCommandlet {
 
     super(context);
     this.newProject = add(new StringProperty("", true, "project"));
-    this.skipRepositories = add(new FlagProperty("--skip-repositories"));
     this.codeRepositoryFlag = add(new FlagProperty("--code"));
     add(this.settingsRepo);
   }
@@ -66,14 +61,7 @@ public class CreateCommandlet extends AbstractUpdateCommandlet {
     initializeProject(newProjectPath);
     this.context.setIdeHome(newProjectPath);
     super.run();
-
-    if (this.skipRepositories.isTrue()) {
-      this.context.info("Skipping the cloning of project repositories as specified by the user.");
-    } else {
-      updateRepositories();
-    }
     this.context.success("Successfully created new project '{}'.", newProjectName);
-
   }
 
   private void initializeCodeRepository(String repoUrl) {
@@ -102,15 +90,10 @@ public class CreateCommandlet extends AbstractUpdateCommandlet {
     fileAccess.mkdirs(newInstancePath.resolve(IdeContext.FOLDER_WORKSPACES).resolve(IdeContext.WORKSPACE_MAIN));
   }
 
-  private void updateRepositories() {
-
-    this.context.getCommandletManager().getCommandlet(RepositoryCommandlet.class).run();
-  }
-
   @Override
   protected void updateSettings() {
 
-    if (codeRepositoryFlag.isTrue()) {
+    if (this.codeRepositoryFlag.isTrue()) {
       String codeRepository = this.settingsRepo.getValue();
       if (codeRepository == null || codeRepository.isBlank()) {
         String message = """
