+++ cli/src/main/java/com/devonfw/tools/ide/commandlet/UpdateCommandlet.java
@@ -22,10 +22,4 @@ public class UpdateCommandlet extends AbstractUpdateCommandlet {
 
     return "update";
   }
-
-  @Override
-  public void run() {
-
-    super.run();
-  }
 }
